+++ backend/controllers/food.js
@@ -4,21 +4,29 @@ import dishModel from "../models/food.js"
 
 // add dish item
 const addDish = async (req, res) => {
-  let image_filename = `${req.file.filename}`
+  const { name, description, price, category, rating } = req.body
+  const { filename } = req.file
+  if (!name || !description || !price || !category || !rating || !filename) {
+    return res.json({
+      error: "Missing required fields.",
+      required: ["name", "description", "price", "category", "rating"],
+    })
+  }
+
   const food = new dishModel({
-    name: req.body.name,
-    description: req.body.description,
-    price: req.body.price,
-    category: req.body.category,
-    rating: req.body.rating,
-    image: image_filename,
+    name: name,
+    description: description,
+    price: price,
+    category: category,
+    rating: rating,
+    image: `${filename}`,
   })
+
   try {
     await food.save()
     res.json({ success: true, message: "Dish Added" })
-  } catch (error) {
-    console.log(error)
-    res.json({ success: false, message: "Error" })
+  } catch (err) {
+    res.json({ success: false, message: `Error in adding dish: ${err}` })
   }
 }
 
@@ -26,40 +34,63 @@ const addDish = async (req, res) => {
 const listDishes = async (req, res) => {
   try {
     const dishes = await dishModel.find({})
+    if (!dishes) {
+      res.json({ success: false, message: "Dishes not found!" })
+      return
+    }
     res.json({ success: true, data: dishes })
-  } catch (error) {
-    console.log(error)
-    res.json({ success: false, message: "Error" })
+  } catch (err) {
+    res.json({ success: false, message: `Error in listing dishes: ${err}` })
   }
 }
 
 // remove dishes
 const removeDish = async (req, res) => {
+  const { id } = req.body
+  if (!id) {
+    return res.json({ success: false, message: "Missing dish's ID field!" })
+  }
+
   try {
-    const dish = await dishModel.findById(req.body.id)
+    const dish = await dishModel.findById(id)
+    if (!dish) {
+      res.json({ success: false, message: "Dish not found!" })
+      return
+    }
     fs.unlink(`uploads/${dish.image}`, () => {})
-    await dishModel.findByIdAndDelete(req.body.id)
+    const deleteDish = await dishModel.findByIdAndDelete(id)
+    if (!deleteDish) {
+      res.json({ success: false, message: "Error in removing the dish!" })
+      return
+    }
     res.json({ success: true, message: "Dish Removed" })
-  } catch (error) {
-    console.log(error)
-    res.json({ success: false, message: "Error" })
+  } catch (err) {
+    res.json({ success: false, message: `Error in removing dish: ${err}` })
   }
 }
 
 // update dish item
 const updateDish = async (req, res) => {
+  const { _id, name, description, price, category, rating } = req.body
+  if (!_id || !name || !description || !price || !category || !rating) {
+    return res.json({
+      error: "Missing required fields.",
+      required: ["userID", "items", "amount", "deliveryType", "address"],
+    })
+  }
+
   try {
-    const dish = await dishModel.findById(req.body._id)
+    const dish = await dishModel.findById(_id)
     if (!dish) {
-      return res.json({ success: false, message: "Dish not found" })
+      return res.json({ success: false, message: "Dish not found!" })
     }
 
     let updatedData = {
-      name: req.body.name || dish.name,
-      description: req.body.description || dish.description,
-      price: req.body.price || dish.price,
-      category: req.body.category || dish.category,
-      rating: req.body.rating || dish.rating,
+      name: name || dish.name,
+      description: description || dish.description,
+      price: price || dish.price,
+      category: category || dish.category,
+      rating: rating || dish.rating,
     }
 
     // Check if an image is uploaded
@@ -70,42 +101,41 @@ const updateDish = async (req, res) => {
       // Remove the old image file if it exists
       if (dish.image) {
         fs.unlink(`uploads/${dish.image}`, (err) => {
-          if (err) console.error("Error removing old image:", err)
+          if (err)
+            res.json({
+              success: false,
+              message: `Error removing old image: ${err}`,
+            })
         })
       }
     }
 
-    await dishModel.findByIdAndUpdate(req.body._id, updatedData, { new: true })
-    res.json({ success: true, message: "Dish updated" })
-  } catch (error) {
-    console.error("Error in updating dish:", error)
-    res.json({ success: false, message: "Error" })
+    const newDish = await dishModel.findByIdAndUpdate(_id, updatedData, {
+      new: true,
+    })
+    if (!newDish) {
+      return res.json({ success: false, message: "Error in updating dish!" })
+    }
+    res.json({ success: true, message: "Dish updated!" })
+  } catch (err) {
+    res.json({ success: false, message: `Error in updating dish: ${err}` })
   }
 }
 
 const getDish = async (req, res) => {
-  const { id } = req.params
-
-  if (!id) {
-    return res
-      .status(400)
-      .json({ success: false, message: "Missing id in the request body." })
+  const { dishID } = req.params
+  if (!dishID) {
+    return res.json({ success: false, message: "Missing dishID field!" })
   }
 
   try {
-    const dish = await dishModel.findOne({ _id: id })
+    const dish = await dishModel.findOne({ _id: dishID })
     if (!dish) {
-      return res
-        .status(404)
-        .json({ success: false, message: "Dish not found." })
+      return res.json({ success: false, message: "Dish not found!" })
     }
     res.json({ success: true, data: dish })
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({
-      success: false,
-      message: "Error occurred while fetching the dish.",
-    })
+  } catch (err) {
+    res.json({ success: false, message: `Error in retrieving dish: ${err}` })
   }
 }
